+++ src/main/java/com/bootcamp/springBootDemo/EmployeeController.java
@@ -12,17 +12,17 @@ import java.util.Map;
 public class EmployeeController {
 
     //    using memory to host the data
-    private final Map<Long, Employee> employees = new HashMap<>(Map.of(
-            1L, new Employee(1L, "John Smith", 32, Gender.MALE, 5000.0),
-            2L, new Employee(2L, "Jane Johnson", 28, Gender.FEMALE, 6000.0),
-            3L, new Employee(3L, "David Williams", 35, Gender.MALE, 5500.0),
-            4L, new Employee(4L, "Emily Brown", 23, Gender.FEMALE, 4500.0),
-            5L, new Employee(5L, "Michael Jones", 40, Gender.MALE, 7000.0)));
+    private final Map<Integer, Employee> employees = new HashMap<>(Map.of(
+            1, new Employee(1, "John Smith", 32, Gender.MALE, 5000.0),
+            2, new Employee(2, "Jane Johnson", 28, Gender.FEMALE, 6000.0),
+            3, new Employee(3, "David Williams", 35, Gender.MALE, 5500.0),
+            4, new Employee(4, "Emily Brown", 23, Gender.FEMALE, 4500.0),
+            5, new Employee(5, "Michael Jones", 40, Gender.MALE, 7000.0)));
 
     @PostMapping
     @ResponseStatus(HttpStatus.CREATED)
     public Employee create(@RequestBody Employee request) {
-        Long id = (long) (employees.size() + 1);
+        int id = employees.size() + 1;
         Employee newEmployee = new Employee(id, request.getName(), request.getAge(), request.getGender(),
                 request.getSalary());
         employees.put(id, newEmployee);
@@ -30,7 +30,7 @@ public class EmployeeController {
     }
 
     @GetMapping("/{id}")
-    public Employee getById(@PathVariable Long id) {
+    public Employee getById(@PathVariable Integer id) {
         return employees.get(id);
     }
